StudyBuddy/requirements.txt
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.6.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
dj-database-url==2.1.0